# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MT-Engine",
    "description": "M-Team \u514d\u8d39\u79cd\u5b50\u730e\u624b",
    "version": "6.6.2"
  },
  "paths": {
...
    "/api/panel/history": {
      "get": {
        "summary": "Get Panel History",
        "description": "\u83b7\u53d6\u6d41\u91cf\u5386\u53f2\u6570\u636e\n\nArgs:\n    range: \u65f6\u95f4\u8303\u56f4\uff0c\u53ef\u9009\u503c: 1h, 6h, 12h, 24h, 7d, 30d\n\nReturns:\n    {\n        \"range\": \"24h\",\n        \"aggregation\": \"5min\",\n        \"data_points\": [...]\n    }",
        "operationId": "get_panel_history_api_panel_history_get",
        "parameters": [
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "1h",
                "6h",
                "12h",
                "24h",
                "7d",
                "30d"
              ],
              "type": "string",
              "default": "24h",
              "title": "Range"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanelHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/panel/share-ratio": {
      "get": {
        "summary": "Get Panel Share Ratio",
        "description": "\u83b7\u53d6\u5206\u4eab\u7387\u5386\u53f2\u6570\u636e\n\nArgs:\n    range: \u65f6\u95f4\u8303\u56f4\uff0c\u53ef\u9009\u503c: 1h, 6h, 12h, 24h, 7d, 30d\n\nReturns:\n    {\n        \"range\": \"24h\",\n        \"data_points\": [...],\n        \"current\": 1.23,\n        \"highest\": 1.50,\n        \"lowest\": 1.00,\n        \"change_24h\": 0.05\n    }",
        "operationId": "get_panel_share_ratio_api_panel_share_ratio_get",
        "parameters": [
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "1h",
                "6h",
                "12h",
                "24h",
                "7d",
                "30d"
              ],
              "type": "string",
              "default": "24h",
              "title": "Range"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanelShareRatioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/panel/torrents": {
      "get": {
        "summary": "Get Panel Torrents",
        "description": "\u83b7\u53d6\u79cd\u5b50\u5217\u8868\uff08\u5e26\u7b5b\u9009\uff09\n\nArgs:\n    tag: \u6807\u7b7e\u7b5b\u9009 (\"\u58f0\u5450\u505a\u79cd\" | \"\u96f7\u8fbe\u4e0b\u8f7d\" | \"PILOT\")\n    status: \u72b6\u6001\u7b5b\u9009 (\"downloading\" | \"seeding\" | \"paused\" | \"completed\")\n\nReturns:\n    {\n        \"torrents\": List[Dict],\n        \"total_count\": int,\n        \"filtered_count\": int\n    }",
        "operationId": "get_panel_torrents_api_panel_torrents_get",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanelTorrentsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/panel/torrents/pause": {
      "post": {
        "summary": "Pause Torrents",
        "description": "\u6279\u91cf\u6682\u505c\u79cd\u5b50",
        "operationId": "pause_torrents_api_panel_torrents_pause_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PauseTorrentsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanelPauseTorrentsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/panel/torrents/resume": {
      "post": {
        "summary": "Resume Torrents",
        "description": "\u6279\u91cf\u6062\u590d\u79cd\u5b50",
        "operationId": "resume_torrents_api_panel_torrents_resume_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeTorrentsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanelResumeTorrentsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/panel/torrents/delete": {
      "post": {
        "summary": "Delete Torrents",
        "description": "\u6279\u91cf\u5220\u9664\u79cd\u5b50",
        "operationId": "delete_torrents_api_panel_torrents_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteTorrentsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanelDeleteTorrentsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeleteTorrentsRequest": {
        "properties": {
          "hashes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hashes"
          },
          "delete_files": {
            "type": "boolean",
            "title": "Delete Files",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "hashes"
        ],
        "title": "DeleteTorrentsRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PanelDeleteTorrentsResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "deleted_count": {
            "type": "integer",
            "title": "Deleted Count",
            "default": 0
          },
          "failed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Failed"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "PanelDeleteTorrentsResponse",
        "description": "Batch delete response"
      },
      "PanelHistoryResponse": {
        "properties": {
          "range": {
            "type": "string",
            "enum": [
              "1h",
              "6h",
              "12h",
              "24h",
              "7d",
              "30d"
            ],
            "title": "Range"
          },
          "aggregation": {
            "type": "string",
            "title": "Aggregation",
            "default": "none"
          },
          "data_points": {
            "items": {
              "$ref": "#/components/schemas/PanelTrafficDataPoint"
            },
            "type": "array",
            "title": "Data Points"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "range"
        ],
        "title": "PanelHistoryResponse",
        "description": "PANEL traffic history response"
      },
      "PanelMetricPoint": {
        "properties": {
          "uploaded": {
            "type": "integer",
            "title": "Uploaded",
            "default": 0
          },
          "downloaded": {
            "type": "integer",
            "title": "Downloaded",
            "default": 0
          }
        },
        "type": "object",
        "title": "PanelMetricPoint",
        "description": "Cumulative upload/download point"
      },
      "PanelPauseTorrentsResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "paused_count": {
            "type": "integer",
            "title": "Paused Count",
            "default": 0
          },
          "failed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Failed"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "PanelPauseTorrentsResponse",
        "description": "Batch pause response"
      },
      "PanelResumeTorrentsResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "resumed_count": {
            "type": "integer",
            "title": "Resumed Count",
            "default": 0
          },
          "failed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Failed"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "PanelResumeTorrentsResponse",
        "description": "Batch resume response"
      },
      "PanelShareRatioDataPoint": {
        "properties": {
          "timestamp": {
            "type": "integer",
            "title": "Timestamp"
          },
          "share_ratio": {
            "type": "number",
            "title": "Share Ratio",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "timestamp"
        ],
        "title": "PanelShareRatioDataPoint",
        "description": "Share-ratio history point"
      },
      "PanelShareRatioResponse": {
        "properties": {
          "range": {
            "type": "string",
            "enum": [
              "1h",
              "6h",
              "12h",
              "24h",
              "7d",
              "30d"
            ],
            "title": "Range"
          },
          "data_points": {
            "items": {
              "$ref": "#/components/schemas/PanelShareRatioDataPoint"
            },
            "type": "array",
            "title": "Data Points"
          },
          "current": {
            "type": "number",
            "title": "Current",
            "default": 0
          },
          "highest": {
            "type": "number",
            "title": "Highest",
            "default": 0
          },
          "lowest": {
            "type": "number",
            "title": "Lowest",
            "default": 0
          },
          "change_24h": {
            "type": "number",
            "title": "Change 24H",
            "default": 0
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "range"
        ],
        "title": "PanelShareRatioResponse",
        "description": "PANEL share-ratio history response"
      },
...
      "PanelTorrentItem": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "hash": {
            "type": "string",
            "title": "Hash",
            "default": ""
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "size": {
            "type": "integer",
            "title": "Size",
            "default": 0
          },
          "size_display": {
            "type": "string",
            "title": "Size Display",
            "default": ""
          },
          "progress": {
            "type": "number",
            "title": "Progress",
            "default": 0
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": ""
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "ratio": {
            "type": "number",
            "title": "Ratio",
            "default": 0
          },
          "uploaded": {
            "type": "integer",
            "title": "Uploaded",
            "default": 0
          },
          "downloaded": {
            "type": "integer",
            "title": "Downloaded",
            "default": 0
          },
          "upload_speed": {
            "type": "integer",
            "title": "Upload Speed",
            "default": 0
          },
          "download_speed": {
            "type": "integer",
            "title": "Download Speed",
            "default": 0
          },
          "seeders": {
            "type": "integer",
            "title": "Seeders",
            "default": 0
          },
          "leechers": {
            "type": "integer",
            "title": "Leechers",
            "default": 0
          },
          "added_on": {
            "type": "integer",
            "title": "Added On",
            "default": 0
          },
          "eta": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eta"
          },
          "health": {
            "$ref": "#/components/schemas/PanelTorrentHealth"
          },
          "mteam_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mteam Id"
          },
          "quality_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Metadata"
          }
        },
        "type": "object",
        "title": "PanelTorrentItem",
        "description": "qBittorrent torrent item displayed by PANEL"
      },
      "PanelTorrentsResponse": {
        "properties": {
          "torrents": {
            "items": {
              "$ref": "#/components/schemas/PanelTorrentItem"
            },
            "type": "array",
            "title": "Torrents"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count",
            "default": 0
          },
          "filtered_count": {
            "type": "integer",
            "title": "Filtered Count",
            "default": 0
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "title": "PanelTorrentsResponse",
        "description": "PANEL torrent list response"
      },
      "PanelTrafficDataPoint": {
        "properties": {
          "timestamp": {
            "type": "integer",
            "title": "Timestamp"
          },
          "mteam": {
            "$ref": "#/components/schemas/PanelMetricPoint"
          },
          "qbittorrent": {
            "$ref": "#/components/schemas/PanelMetricPoint"
          }
        },
        "type": "object",
        "required": [
          "timestamp"
        ],
        "title": "PanelTrafficDataPoint",
        "description": "Traffic history point"
      },
...
      "PauseTorrentsRequest": {
        "properties": {
          "hashes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hashes"
          }
        },
        "type": "object",
        "required": [
          "hashes"
        ],
        "title": "PauseTorrentsRequest"
      },
      "ResumeTorrentsRequest": {
        "properties": {
          "hashes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hashes"
          }
        },
        "type": "object",
        "required": [
          "hashes"
        ],
        "title": "ResumeTorrentsRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "description": "MT Engine API key sent as an Authorization bearer token.",
        "scheme": "bearer"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "description": "MT Engine API key sent in the X-MT-ENGINE-Key header.",
        "in": "header",
        "name": "X-MT-ENGINE-Key"
      }
    }
  }
}